requirements.txt
# sudo pip install -f requirements.txt
django
django-ratings
django-piston
xlrd
